refactor: clean code and update tests
