+++ src/cadence/spack.config.js
@@ -2,7 +2,7 @@ const { config } = require('@swc/core/spack')
 
 module.exports = config({
   entry: {
-    web: __dirname + '/src/cadence2/index.ts',
+    web: __dirname + '/src/index.ts',
   },
   output: {
     path: __dirname + '/../playground/public',
