--- src/cadence/src/cadence2/index.ts
@@ -1,8 +1,0 @@
-import { loadSample } from './loadSample'
-import { loop } from './loop'
-import { initCadence } from './init'
-import type { InitResult, PreloadResult, Preload } from './init'
-import { wait } from './wait'
-
-export { loadSample, loop, initCadence, wait }
-export type { InitResult, PreloadResult, Preload }
