+++ src/cadence/src/init/index.ts
@@ -1,0 +1,3 @@
+export { initCadence } from './initCadence'
+export { loadSample } from './loadSample'
+export type { InitResult, PreloadResult, Preload } from './initCadence'
