+++ src/cadence/src/play/index.ts
@@ -1,0 +1,7 @@
+export { loop } from './loop'
+export { playSample } from './playSample'
+export { wait } from './wait'
+export type { PlayerOptions, SampleNode } from './playSample'
+export type { Effect, Time } from './types'
+export type { Distorsion } from './distorsion'
+export type { Fade, FadeIn, FadeOut } from './fades'
