--- src/cadence/src/v1/index.ts
@@ -1,2 +1,0 @@
-export * from './types'
-export { Cadence } from './Cadence'
